requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
Flask-CORS
Flask-WTF
WTForms
python-dotenv
Werkzeug
email-validator
Pillow
geopy
requests
pandas
openpyxl
google-generativeai
numpy
opencv-python-headless
thefuzz
feedparser
easyocr
reportlab
gunicorn
psycopg2-binary
